[WIP] libgenpattern. Seamless!

+++ libgenpattern/libgenpattern/include/suitability.h
@@ -8,9 +8,9 @@ typedef struct {
   GPPolygon* polygons_buffer;
   gp_float target;
   GPPolygon* polygons;
-  int32_t polygons_len;
+  size_t polygons_len;
   GPPolygon* collection;
-  int32_t collection_len;
+  size_t collection_len;
   GPPolygon* canvas;
   GPPolygon* canvas_outside_areas;
   GPPolygon* ref;
